Refactor : Answer, Comment( memberId 제거 및 JWT관련 로직 수정)

+++ backend/stackOverFlowClone/src/main/java/com/preproject/stackOverFlowClone/comment/service/CommentService.java
@@ -1,6 +1,7 @@
 package com.preproject.stackOverFlowClone.comment.service;
 
-import com.preproject.stackOverFlowClone.answer.entity.Answer;
+import com.preproject.stackOverFlowClone.ask.entity.Ask;
+import com.preproject.stackOverFlowClone.auth.utils.GetAuthUserUtils;
 import com.preproject.stackOverFlowClone.comment.dto.CommentResponseDto;
 import com.preproject.stackOverFlowClone.comment.dto.CommentSaveDto;
 import com.preproject.stackOverFlowClone.comment.dto.CommentUpdateDto;
@@ -10,39 +11,50 @@ import com.preproject.stackOverFlowClone.dto.MultiResponseDto;
 import com.preproject.stackOverFlowClone.dto.SingleResponseDto;
 import com.preproject.stackOverFlowClone.exception.BusinessLogicException;
 import com.preproject.stackOverFlowClone.exception.ExceptionCode;
+import com.preproject.stackOverFlowClone.member.entity.Member;
+import com.preproject.stackOverFlowClone.member.repository.MemberRepository;
+import com.preproject.stackOverFlowClone.member.service.MemberService;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDateTime;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 public class CommentService {
-
+    private final MemberRepository memberRepository;
     private final CommentRepository commentRepository;
+    private final MemberService memberService;
 
-    public CommentService(CommentRepository commentRepository) {
+    public CommentService(MemberRepository memberRepository,
+                          CommentRepository commentRepository,
+                          MemberService memberService) {
+        this.memberRepository = memberRepository;
         this.commentRepository = commentRepository;
+        this.memberService = memberService;
     }
 
 
     public void saveComment(CommentSaveDto commentSaveDto) {
-
-        boolean validMember = commentRepository.existsByMemberId(commentSaveDto.getMemberId());
+//        boolean validMember = commentRepository.existsByMemberId(commentSaveDto.getMemberId());
         boolean validAsk = commentRepository.existsByAskId(commentSaveDto.getAskId());
         boolean validAnswer = commentRepository.existsByAnswerId(commentSaveDto.getAnswerId());
 
-        if(!validMember || !validAsk || !validAnswer) {
+        if(!validAsk || !validAnswer) {
             throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
         }
+        else {
+            Member member = memberService.getLoginMember();
+            Comment comment = Comment.of(commentSaveDto);
+            comment.setMemberId(member.getId());
 
-
-        Comment comment = Comment.of(commentSaveDto);
-
-        commentRepository.save(comment);
+            commentRepository.save(comment);
+        }
     }
 
     @Transactional
@@ -50,18 +62,34 @@ public class CommentService {
 
         Comment findComment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND));
+        Member loginmember = memberService.getLoginMember();
 
-        findComment.update(commentUpdateDto);
+        if (loginmember.getId().equals(findComment.getMemberId())) {
+            findComment.setContent(commentUpdateDto.getContent());
+            findComment.setCreatedAt(LocalDateTime.now());
 
-        String memberName = commentRepository.findMemberNameByMemberId(findComment.getMemberId())
-                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));;
-        return CommentResponseDto.of(memberName, findComment);
+            findComment.update(commentUpdateDto);
+
+            String memberName = commentRepository.findMemberNameByMemberId(findComment.getMemberId())
+                    .orElseThrow(() -> new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND));;
+            return CommentResponseDto.of(memberName, findComment);
+        } else {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCH);
+        }
     }
 
     public void deleteComment(Long commentId) {
+        Member loginMember = memberService.getLoginMember();
+        Optional<Comment> findComment = Optional.ofNullable(commentRepository.findById(commentId)
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.COMMENT_NOT_FOUND)));
+        Comment comment = findComment.get();
 
-        commentRepository.deleteById(commentId);
-
+        if (loginMember.getId().equals(comment.getMemberId())) {
+            commentRepository.deleteById(commentId);
+        }
+        else{
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCH);
+        }
     }
 
     public SingleResponseDto<CommentResponseDto> findComment(Long commentId) {
